% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{calculo}
\alias{calculo}
\title{Calcula diferencias estadísticas por nodo entre dos matrices de correlación}
\usage{
calculo(matriz1, matriz2, nodo)

calculo(matriz1, matriz2, nodo)
}
\arguments{
\item{matriz1}{Primera matriz de correlación.}

\item{matriz2}{Segunda matriz de correlación.}

\item{nodo}{Nombre del nodo para el cual se calcularán las diferencias.}
}
\value{
Data frame con la diferencia media y el valor p del test t.

Data frame con la diferencia media y el valor p del test t.
}
\description{
Calcula diferencias estadísticas por nodo entre dos matrices de correlación

Calcula diferencias estadísticas por nodo entre dos matrices de correlación
}
