% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes Dataset}
\format{
A data frame with the following variables:
\describe{
  \item{Pregnancies}{Number of times pregnant}
  \item{Glucose}{Plasma glucose concentration a 2 hours in an oral glucose tolerance test}
  \item{BloodPressure}{Diastolic blood pressure (mm Hg)}
  \item{SkinThickness}{Triceps skin fold thickness (mm)}
  \item{Insulin}{2-Hour serum insulin (mu U/ml)}
  \item{BMI}{Body mass index (weight in kg/(height in m)^2)}
  \item{DiabetesPedigreeFunction}{Diabetes pedigree function}
  \item{Age}{Age (years)}
  \item{Outcome}{Class variable (0 or 1)}
}

A data frame with rows and columns:
\describe{
  \item{Pregnancies}{Number of times pregnant}
  \item{Glucose}{Plasma glucose concentration a 2 hours in an oral glucose tolerance test}
  \item{BloodPressure}{Diastolic blood pressure (mm Hg)}
  \item{SkinThickness}{Triceps skin fold thickness (mm)}
  \item{Insulin}{2-Hour serum insulin (mu U/ml)}
  \item{BMI}{Body mass index (weight in kg/(height in m)^2)}
  \item{DiabetesPedigreeFunction}{Diabetes pedigree function}
  \item{Age}{Age (years)}
  \item{Outcome}{Class variable (0 or 1)}
}
}
\usage{
data(diabetes)

data(diabetes)
}
\description{
This dataset contains information about diabetes patients.

This dataset contains information about diabetes patients.
}
\keyword{datasets}
