% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{divide_y_elimina}
\alias{divide_y_elimina}
\title{Divide y elimina la variable de respuesta}
\usage{
divide_y_elimina(df, respuesta_var)

divide_y_elimina(df, respuesta_var)
}
\arguments{
\item{df}{Data frame con los datos.}

\item{respuesta_var}{Nombre de la variable de respuesta.}
}
\value{
Una lista con dos data frames: uno para cada valor de la variable de respuesta.

Una lista con dos data frames: uno para cada valor de la variable de respuesta.
}
\description{
Divide y elimina la variable de respuesta

Divide y elimina la variable de respuesta
}
