% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{matriz_correlacion}
\alias{matriz_correlacion}
\title{Calcula la matriz de correlación parcial}
\usage{
matriz_correlacion(df, nivel_significacion = 0.05)

matriz_correlacion(df, nivel_significacion = 0.05)
}
\arguments{
\item{df}{Data frame con los datos.}

\item{nivel_significacion}{Nivel de significación para los valores p (por defecto 0.05).}
}
\value{
Matriz de correlación parcial ajustada.

Matriz de correlación parcial ajustada.
}
\description{
Calcula la matriz de correlación parcial

Calcula la matriz de correlación parcial
}
