% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{test}
\alias{test}
\title{Realiza el test de permutaciones para la diferencia total de matrices de correlación}
\usage{
test(df, respuesta_var, n_permutaciones = 1000, nivel_significacion = 0.05)

test(df, respuesta_var, n_permutaciones = 1000, nivel_significacion = 0.05)
}
\arguments{
\item{df}{Data frame con los datos.}

\item{respuesta_var}{Nombre de la variable de respuesta.}

\item{n_permutaciones}{Número de permutaciones (por defecto 1000).}

\item{nivel_significacion}{Nivel de significación para los valores p (por defecto 0.05).}
}
\value{
Vector de diferencias totales de las matrices de correlación para cada permutación.

Vector de diferencias totales de las matrices de correlación para cada permutación.
}
\description{
Realiza el test de permutaciones para la diferencia total de matrices de correlación

Realiza el test de permutaciones para la diferencia total de matrices de correlación
}
